Homework 2 finished

+++ Homework2/1/script1.py
@@ -3,7 +3,7 @@ import numpy as np
 
 
 def convolution(a, b):  # Convolving two signals
-    a = np.concatenate([ np.zeros( len(b) / 2 - 1 ), a,  np.zeros( len(b) / 2 )])
+    a = np.concatenate([ np.zeros( (int)(len(b) / 2 - 1) ), a,  np.zeros( (int)(len(b) / 2) )])
     b = b[::-1]
 
     step = len(b)
@@ -29,10 +29,10 @@ t = np.arange(start, end, d_step)
 
 # DT signals
 x1 = [1, 2, 3, 2, 1, 2, 3]
-x1 = np.concatenate([np.zeros(len(t) / 2 - len(x1) / 2 - 1), x1, np.zeros(len(t) / 2 - len (x1) / 2)])
+x1 = np.concatenate([np.zeros( (int)(len(t) / 2) - (int)(len(x1) / 2) - 1), x1, np.zeros((int)(len(t) / 2) - (int)(len (x1) / 2))])
 
 x2 = [2, 4, -2]
-x2 = np.concatenate([np.zeros(len(t) / 2 - len(x2) / 2 - 1), x2, np.zeros(len(t) / 2 - len (x2) / 2)])
+x2 = np.concatenate([np.zeros( (int)(len(t) / 2) - (int)(len(x2) / 2) - 1), x2, np.zeros((int)(len(t) / 2) - (int)(len (x2) / 2))])
 
 x3 = convolution(x1, x2)
